requirements.txt
Django==5.1.4
django-allauth==65.3.0
python-decouple==3.8
Pillow==11.1.0
google-generativeai==0.8.4
huggingface_hub
python-docx==1.1.2
gunicorn==23.0.0
whitenoise==6.8.2
django-ratelimit==4.1.0
PyPDF2==3.0.1
